service: select project vpc
Project VPC's can be specified for service create and update.

`avn service create --cloud <cloud name> --project-vpc-id <UUID>` will
place the service into a project VPC. Note that the VPC has to be in the
same cloud as the service.

`avn service create --no-project-vpc` puts the service into a public
network in the selected cloud, avoiding the default behavior where the
service would be put into a project VPC if one exists for the project in
that cloud. If there is no VPC in the selected cloud, the argument has
no effect.

The arguments have the same meaning for `avn service update`, except
that omitting the arguments will leave the service in the network it
currently is in (unless a cloud migration is initiated using `--cloud`).