Support easier read replica creation and upcoming API
`service create` command supports --read-replica-for parameter, which
can be used to indicate that the service being created should be a
read replica of the given service. This automatically sets the
appropriate user configuration options for PostgreSQL services and for
other services that use the upcoming read_replica service integration
API that is defined instead.