Add uniform support for data output format.
Modifies argx to collect a uniform set of specifiers at launch for log
level, output format and output stream. This is passed to all command-line
utilities and affects output via the print_response function.

This centralizes output selection and enables all commands to emit output
using any supported format.

print_response is shimmed to allow this version of the library to be
backwards compatible, pending removal of redundant parameters from
other tools.

This commit adds a dependency of ruamel.yaml as part of supporting YAML
as an output format.