Deprecate / remove service group concept
The service group concept has been removed on backend side. The API
still accepts and returns the related fields but the input is always
ignored and output is hardcoded dummy value. Remove it also from the
cli except for service create and update commands so that backwards
compatibility is kept.

Note that the Python library change is backwards incompatible as it
was not possible to make the parameter optional while retaining
compatibility.