github: updates following master branch rename to main
`main` is a more inclusive name, and has been the default on github for a
while so let's switch to using that.  Add a workflow to keep `main` &
`master` branches in sync for a while to make sure other, external,
workflows don't break because of this.

That workflow and the `master` branch should be deleted after a couple of
weeks or so.