Fix API client compliance with HTTP/1.1
In HTTP/1.1 Content-Length is not always required to be sent by the
server - if chunked transfer encoding is used, the Content-Length header
can be omitted, and length indicated as part of each chunk. Chunked
transfer encoding is always acceptable in HTTP/1.1 [1] so by effectively
requiring a Content-Length header we are not following the spec.

[1]: https://datatracker.ietf.org/doc/html/rfc9112#section-7.4-2