Add commands to manage permissions
Add commands to configure and list permissions associated to a given
resource.

Note the `permissions update` command operate in `set` mode, so it will
always override the total of the configured permissions.  Also, it is
not possible to work with permissions on project and
organization level on the same command.